LayoutTests/media/video-no-autoplay.html

<video src=content/test.mp4></video>
<script src=video-test.js></script>
<script>
test("video.paused");
waitForEventAndTest("play", "false");
waitForEvent("load", function () { setTimeout(testAndEnd('video.paused', 500)) });
</script>
