LayoutTests/media/video-no-autoplay-expected.txt

TEST(video.paused) OK
EVENT(load)
TEST(video.paused) OK
END OF TEST
